# docs/swagger.yml
swagger: '2.0'
info:
  title: backend.log
  description: |
    Тестовое задание для реализации проекта "Форумы" на курсе по базам данных в
    Технопарке VK (https://park.vk.company).
  version: "0.1.0"
schemes:
  - http
  - https
basePath: /api
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /user/{nickname}/create:
    post:
      summary: Создание нового пользователя
      description: |
        Создание нового пользователя в базе данных.
      operationId: userCreate
      parameters:
        - name: nickname
          in: path
          description: Идентификатор пользователя.
          required: true
          type: string
        - name: profile
          in: body
          description: Данные пользовательского профиля.
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        201:
          description: |
            Пользователь успешно создан.
            Возвращает данные созданного пользователя.
          schema:
            $ref: '#/definitions/User'
        409:
          description: |
            Пользователь уже присутсвует в базе данных.
            Возвращает данные ранее созданных пользователей с тем же nickname-ом иои email-ом.
          schema:
            $ref: '#/definitions/Users'
  /user/{nickname}/profile:
    get:
      summary: Получение информации о пользователе
      description: |
        Получение информации о пользователе форума по его имени.
      consumes: [ ]
      operationId: userGetOne
      parameters:
        - name: nickname
          in: path
          description: Идентификатор пользователя.
          required: true
          type: string
      responses:
        200:
          description: |
            Информация о пользователе.
          schema:
            $ref: '#/definitions/User'
        404:
          description: |
            Пользователь отсутсвует в системе.
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Изменение данных о пользователе
      description: |
        Изменение информации в профиле пользователя.
      operationId: userUpdate
      parameters:
        - name: nickname
          in: path
          description: Идентификатор пользователя.
          required: true
          type: string
        - name: profile
          in: body
          description: Изменения профиля пользователя.
          required: true
          schema:
            $ref: '#/definitions/UserUpdate'
      responses:
        200:
          description: |
            Актуальная информация о пользователе после изменения профиля.
          schema:
            $ref: '#/definitions/User'
        404:
          description: |
            Пользователь отсутсвует в системе.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: |
            Новые данные профиля пользователя конфликтуют с имеющимися пользователями.
          schema:
            $ref: '#/definitions/Error'
definitions:
  Error:
    type: object
    properties:
      message:
        type: string
        readOnly: true
        description: |
          Текстовое описание ошибки.
          В процессе проверки API никаких проверок на содерижимое данного описание не делается.
        example: |
          Can't find user with id #42
...
  User:
    description: |
      Информация о пользователе.
    type: object
    properties:
      nickname:
        type: string
        format: identity
        readOnly: true
        description: |
          Имя пользователя (уникальное поле).
          Данное поле допускает только латиницу, цифры и знак подчеркивания.
          Сравнение имени регистронезависимо.
        example: j.sparrow
      fullname:
        type: string
        description: Полное имя пользователя.
        example: Captain Jack Sparrow
        x-isnullable: false
      about:
        type: string
        format: text
        description: Описание пользователя.
        example: This is the day you will always remember as the day that you almost caught Captain Jack Sparrow!
      email:
        type: string
        format: email
        description: Почтовый адрес пользователя (уникальное поле).
        example: captaina@blackpearl.sea
        x-isnullable: false
    required:
      - fullname
      - email
  Users:
    type: array
    items:
      $ref: '#/definitions/User'
  UserUpdate:
    description: |
      Информация о пользователе.
    type: object
    properties:
      fullname:
        type: string
        description: Полное имя пользователя.
        example: Captain Jack Sparrow
      about:
        type: string
        format: text
        description: Описание пользователя.
        example: This is the day you will always remember as the day that you almost caught Captain Jack Sparrow!
      email:
        type: string
        format: email
        description: Почтовый адрес пользователя (уникальное поле).
        example: captaina@blackpearl.sea
...